refactor: move hostnames, routes and users from stats to context
