+++ Aikido.Zen.Test/StatsTests.cs
@@ -1,4 +1,3 @@
-
 using Aikido.Zen.Core;
 using Aikido.Zen.Core.Models;
 
@@ -516,225 +515,6 @@ namespace Aikido.Zen.Test
                 // Optional: Force compression and check results if needed, but adds complexity.
                 // stats.ForceCompress();
                 // // Add assertions based on compressed data if required
-            });
-        }
-
-        [Test]
-        public void AddHostname_WithValidHostname_AddsToHostnames()
-        {
-            // Arrange
-            var stats = new AgentStats();
-            var hostname = "example.com:8080";
-
-            // Act
-            stats.AddHostname(hostname);
-
-            // Assert
-            var hostnames = stats.Hostnames;
-            Assert.That(hostnames.Count(), Is.EqualTo(1));
-            var host = hostnames.First();
-            Assert.That(host.Hostname, Is.EqualTo("example.com"));
-            Assert.That(host.Port, Is.EqualTo(8080));
-        }
-
-        [Test]
-        public void AddHostname_WithInvalidHostname_DoesNotAddToHostnames()
-        {
-            // Arrange
-            var stats = new AgentStats();
-            var hostname = "";
-
-            // Act
-            stats.AddHostname(hostname);
-
-            // Assert
-            Assert.That(stats.Hostnames, Is.Empty);
-        }
-
-        [Test]
-        public void AddUser_WithValidUser_AddsToUsers()
-        {
-            // Arrange
-            var stats = new AgentStats();
-            var user = new User("123", "Test User");
-            var ipAddress = "192.168.1.1";
-
-            // Act
-            stats.AddUser(user, ipAddress);
-
-            // Assert
-            var users = stats.Users;
-            Assert.That(users?.Count(), Is.EqualTo(1));
-            var addedUser = users.First();
-            Assert.That(addedUser.Id, Is.EqualTo("123"));
-            Assert.That(addedUser.Name, Is.EqualTo("Test User"));
-            Assert.That(addedUser.LastIpAddress, Is.EqualTo(ipAddress));
-            Assert.That(addedUser.LastSeenAt, Is.GreaterThan(0));
-        }
-
-        [Test]
-        public void AddUser_WithInvalidUser_DoesNotAddToUsers()
-        {
-            // Arrange
-            var stats = new AgentStats();
-            User user = null;
-            var ipAddress = "192.168.1.1";
-
-            // Act
-            stats.AddUser(user, ipAddress);
-
-            // Assert
-            Assert.That(stats.Users, Is.Empty);
-        }
-
-        [Test]
-        public void AddRoute_WithValidContext_AddsToRoutes()
-        {
-            // Arrange
-            var stats = new AgentStats();
-            var context = new Aikido.Zen.Core.Context
-            {
-                Route = "/api/test",
-                Method = "GET"
-            };
-
-            // Act
-            stats.AddRoute(context);
-
-            // Assert
-            var routes = stats.Routes;
-            Assert.That(routes.Count(), Is.EqualTo(1));
-            var route = routes.First();
-            Assert.That(route.Path, Is.EqualTo("/api/test"));
-            Assert.That(route.Method, Is.EqualTo("GET"));
-        }
-
-        [Test]
-        public void AddRoute_WithInvalidContext_DoesNotAddToRoutes()
-        {
-            // Arrange
-            var stats = new AgentStats();
-            Aikido.Zen.Core.Context context = null;
-
-            // Act
-            stats.AddRoute(context);
-
-            // Assert
-            Assert.That(stats.Routes, Is.Empty);
-        }
-
-        [Test]
-        public void AddHostname_ShouldEvictLeastFrequentlyUsed_WhenMaxReached()
-        {
-            // Arrange
-            const int MaxHostnames = 2000; // Use a smaller number for faster testing maybe?
-            var stats = new AgentStats();
-            var firstHostname = "host0.com:80";
-            stats.Reset();
-
-            // Act
-            // Add the first hostname (Hits = 1)
-            stats.AddHostname(firstHostname);
-
-            // Add MaxHostnames more hostnames, ensuring they all end up with Hits = 2
-            for (int i = 1; i <= MaxHostnames; i++)
-            {
-                var hostname = $"host{i}.com:80";
-                // AddHostname calls AddOrUpdate internally.
-                // When adding host[MaxHostnames], Size >= MaxHostnames triggers eviction *before* add.
-                // LFU is host0 (Hits=1). host0 is evicted.
-                stats.AddHostname(hostname); // Adds host{i}(1). Size reaches MaxHostnames.
-                stats.AddHostname(hostname); // Updates host{i}(2). Size remains MaxHostnames.
-            }
-            // Final state: host1..host[MaxHostnames] (all Hits=2). Size = MaxHostnames. host0 was evicted.
-
-            // Assert
-            Assert.Multiple(() =>
-            {
-                Assert.That(stats.Hostnames.Count(), Is.EqualTo(MaxHostnames), "Dictionary size should be at max capacity.");
-                // host0 (Hits=1) should have been evicted when host[MaxHostnames] was added.
-                Assert.That(stats.Hostnames.Any(h => h.Hostname == "host0.com"), Is.False, "First hostname (host0) should be evicted as LFU.");
-                // Verify one of the later added hostnames (host1) is still present.
-                Assert.That(stats.Hostnames.Any(h => h.Hostname == "host1.com"), Is.True, "A hostname with higher hits (host1) should remain.");
-                // Verify the last hostname added (host[MaxHostnames]) is present.
-                Assert.That(stats.Hostnames.Any(h => h.Hostname == $"host{MaxHostnames}.com"), Is.True, "The last hostname added should remain.");
-            });
-        }
-
-        [Test]
-        public void AddUser_ShouldEvictLeastFrequentlyUsed_WhenMaxReached()
-        {
-            // Arrange
-            const int MaxUsers = 2000;
-            var stats = new AgentStats();
-            var firstUser = new User("user0", "User 0");
-            var ipAddress = "192.168.0.1";
-            stats.Reset();
-
-            // Act
-            // Add the first user (Hits = 1)
-            stats.AddUser(firstUser, ipAddress);
-
-            // Add MaxUsers more users, ensuring they all end up with Hits = 2
-            for (int i = 1; i <= MaxUsers; i++)
-            {
-                var user = new User($"user{i}", $"User {i}");
-                // AddUser calls AddOrUpdate internally.
-                // When adding user[MaxUsers], Size >= MaxUsers triggers eviction *before* add.
-                // LFU is user0 (Hits=1). user0 is evicted.
-                stats.AddUser(user, ipAddress); // Adds user{i}(1). Size reaches MaxUsers.
-                stats.AddUser(user, ipAddress); // Updates user{i}(2). Size remains MaxUsers.
-            }
-            // Final state: user1..user[MaxUsers] (all Hits=2). Size = MaxUsers. user0 was evicted.
-
-            // Assert
-            Assert.Multiple(() =>
-            {
-                Assert.That(stats.Users.Count(), Is.EqualTo(MaxUsers), "Dictionary size should be at max capacity.");
-                // user0 (Hits=1) should have been evicted when user[MaxUsers] was added.
-                Assert.That(stats.Users.Any(u => u.Id == "user0"), Is.False, "First user (user0) should be evicted as LFU.");
-                // Verify one of the later added users (user1) is still present.
-                Assert.That(stats.Users.Any(u => u.Id == "user1"), Is.True, "A user with higher hits (user1) should remain.");
-                // Verify the last user added (user[MaxUsers]) is present.
-                Assert.That(stats.Users.Any(u => u.Id == $"user{MaxUsers}"), Is.True, "The last user added should remain.");
-            });
-        }
-
-        [Test]
-        public void AddRoute_ShouldEvictLeastFrequentlyUsed_WhenMaxReached()
-        {
-            // Arrange
-            const int MaxRoutes = 5000;
-            var stats = new AgentStats();
-            var firstRouteContext = new Context { Url = "/route0", Method = "GET", Route = "/route0" };
-            stats.Reset();
-
-            // Act
-            // Add the first route (Hits = 1)
-            stats.AddRoute(firstRouteContext);
-
-            // Add MaxRoutes more routes, ensuring they all end up with Hits = 2
-            for (int i = 1; i <= MaxRoutes; i++)
-            {
-                var routeContext = new Context { Url = $"/route{i}", Method = "GET", Route = $"/route{i}" };
-                // AddRoute calls AddOrUpdate internally.
-                // When adding route[MaxRoutes], Size >= MaxRoutes triggers eviction *before* add.
-                // LFU is route0 (Hits=1). route0 is evicted.
-                stats.AddRoute(routeContext); // Adds route{i}(1). Size reaches MaxRoutes.
-                stats.AddRoute(routeContext); // Updates route{i}(2). Size remains MaxRoutes.
-            }
-            // Final state: route1..route[MaxRoutes] (all Hits=2). Size = MaxRoutes. route0 was evicted.
-
-            // Assert
-            Assert.Multiple(() =>
-            {
-                Assert.That(stats.Routes.Count(), Is.EqualTo(MaxRoutes), "Dictionary size should be at max capacity.");
-                // route0 (Hits=1) should have been evicted when route[MaxRoutes] was added.
-                Assert.That(stats.Routes.Any(r => r.Path == "/route0"), Is.False, "First route (route0) should be evicted as LFU.");
-                // Verify one of the later added routes (route1) is still present.
-                Assert.That(stats.Routes.Any(r => r.Path == "/route1"), Is.True, "A route with higher hits (/route1) should remain.");
-                // Verify the last route added (route[MaxRoutes]) is present.
-                Assert.That(stats.Routes.Any(r => r.Path == $"/route{MaxRoutes}"), Is.True, "The last route added should remain.");
             });
         }
     }
